+++ src/main/java/com/hhplus/lecture/application/LectureService.java
@@ -6,10 +6,10 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.hhplus.lecture.domian.Lecture;
-import com.hhplus.lecture.domian.LectureEnrollment;
-import com.hhplus.lecture.domian.repository.LectureEnrollmentRepository;
-import com.hhplus.lecture.domian.repository.LectureRepository;
+import com.hhplus.lecture.domain.Lecture;
+import com.hhplus.lecture.domain.LectureEnrollment;
+import com.hhplus.lecture.domain.repository.LectureEnrollmentRepository;
+import com.hhplus.lecture.domain.repository.LectureRepository;
 import com.hhplus.lecture.interfaces.exception.ErrorMessage;
 
 import lombok.RequiredArgsConstructor;
