+++ src/main/java/com/hhplus/lecture/domain/repository/LectureEnrollmentRepository.java
@@ -1,11 +1,11 @@
-package com.hhplus.lecture.domian.repository;
+package com.hhplus.lecture.domain.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.hhplus.lecture.domian.LectureEnrollment;
+import com.hhplus.lecture.domain.LectureEnrollment;
 
 @Repository
 public interface LectureEnrollmentRepository extends JpaRepository<LectureEnrollment, Long> {
