+++ src/main/java/com/hhplus/lecture/domain/repository/LectureRepository.java
@@ -1,9 +1,9 @@
-package com.hhplus.lecture.domian.repository;
+package com.hhplus.lecture.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.hhplus.lecture.domian.Lecture;
+import com.hhplus.lecture.domain.Lecture;
 
 @Repository
 public interface LectureRepository extends JpaRepository<Lecture, Long> {
